modify ws-example.py config: add 'verbose': True
The config dict passed to the python WebsocketConnection constructor
now requires a 'verbose' key. This commit adds the entry
`'verbose': True` to the config dict that is passed to the
WebsocketConnection constructor in the examples/py/ws-example.py file,
eliminating a 'KeyError' exception that was being thrown by the
call to ws.connect() in that file.